+++ packages/chain-consumer/src/types.ts
@@ -1,0 +1,13 @@
+import {
+  Validator as GrpcValidator,
+  Delegation as GrpcDelegation,
+  UnbondingDelegation as GrpcUnbondingDelegation,
+  DelegationResponse as GrpcDelegationResponse,
+} from '@injectivelabs/chain-api/cosmos/staking/v1beta1/staking_pb'
+
+export {
+  GrpcValidator,
+  GrpcDelegation,
+  GrpcUnbondingDelegation,
+  GrpcDelegationResponse,
+}