+++ packages/web3-strategy/src/types/index.ts
@@ -1,5 +1,6 @@
 import Eip1993Provider from 'eip1193-provider'
 import { ChainId } from '@injectivelabs/ts-types'
+import { Wallet } from './enums'
 
 export * from './enums'
 export * from './strategy'
@@ -17,4 +18,10 @@ export interface ConcreteStrategyOptions {
   rpcUrls: Record<ChainId, string>
   wsRpcUrls: Record<ChainId, string>
   pollingInterval: number
+}
+
+export interface Web3StrategyArguments {
+  chainId: ChainId
+  wallet: Wallet
+  options: ConcreteStrategyOptions
 }