+++ packages/contracts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.17](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@0.0.16...@injectivelabs/contracts@0.0.17) (2021-03-04)
+
+**Note:** Version bump only for package @injectivelabs/contracts
+
+
+
+
+
 ## [0.0.16](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@0.0.15...@injectivelabs/contracts@0.0.16) (2021-03-03)
 
 **Note:** Version bump only for package @injectivelabs/contracts